metasite
This is just a general draft of the homepage of the metasite. I wasn't sure what to do with the layout for the Issues and Our Solution section, so I just made it into a 3 column layout for now. Needs improvement for sure.  I'll brainstorm some more to improve the layout!